Rename utility functions for clarity, fix imports
