+++ raiden_contracts/tests/fixtures/base/address.py
@@ -4,8 +4,8 @@ from typing import Callable
 import pytest
 from eth_utils import denoms, is_address
 
-from raiden_contracts.tests.utils.config import MAX_UINT256
-from raiden_contracts.utils.sign_utils import private_key_to_address
+from raiden_contracts.tests.utils.constants import MAX_UINT256
+from raiden_contracts.utils.sign import private_key_to_address
 
 
 @pytest.fixture
