+++ raiden_contracts/tests/test_channel_close.py
@@ -7,13 +7,13 @@ from raiden_contracts.constants import (
     ChannelState,
 )
 from raiden_contracts.utils.events import check_channel_closed
-from raiden_contracts.tests.utils import ChannelValues
-from raiden_contracts.tests.utils.config import (
+from raiden_contracts.tests.utils import (
     EMPTY_BALANCE_HASH,
     EMPTY_LOCKSROOT,
     EMPTY_ADDITIONAL_HASH,
     EMPTY_SIGNATURE,
     fake_bytes,
+    ChannelValues,
 )
