+++ raiden_contracts/tests/test_channel_cooperative_settle.py
@@ -1,7 +1,7 @@
 import pytest
 from eth_tester.exceptions import TransactionFailed
 from raiden_contracts.utils.events import check_channel_settled
-from raiden_contracts.tests.utils.config import EMPTY_ADDRESS
+from raiden_contracts.tests.utils.constants import EMPTY_ADDRESS
 from raiden_contracts.constants import ChannelEvent
 from web3.exceptions import ValidationError
