+++ raiden_contracts/tests/test_channel_update_transfer.py
@@ -9,14 +9,14 @@ from raiden_contracts.constants import (
     TEST_SETTLE_TIMEOUT_MIN,
 )
 from raiden_contracts.utils.events import check_transfer_updated
-from raiden_contracts.tests.utils import ChannelValues
-from raiden_contracts.tests.utils.config import (
+from raiden_contracts.tests.utils import (
     EMPTY_BALANCE_HASH,
     EMPTY_LOCKSROOT,
     EMPTY_ADDITIONAL_HASH,
     EMPTY_SIGNATURE,
     EMPTY_ADDRESS,
     fake_bytes,
+    ChannelValues,
 )
