+++ raiden_contracts/tests/unit/test_recover_from_signature.py
@@ -1,9 +1,9 @@
 import pytest
 from eth_tester.exceptions import TransactionFailed
 from web3 import Web3
-from raiden_contracts.utils.sign import hash_balance_proof
-from raiden_contracts.tests.utils.config import EMPTY_ADDRESS
-from raiden_contracts.utils.sign_utils import sign
+from raiden_contracts.utils.proofs import hash_balance_proof
+from raiden_contracts.tests.utils.constants import EMPTY_ADDRESS
+from raiden_contracts.utils.sign import sign
 
 
 @pytest.fixture
