+++ raiden_contracts/tests/utils/__init__.py
@@ -1,5 +1,7 @@
 # flake8: noqa
 
-from .config import *
+from .constants import *
+from .mock import *
 from .contracts import *
+from .transaction import *
 from .channel import *
