+++ raiden_contracts/utils/__init__.py
@@ -1,3 +1,9 @@
 # flake8: noqa
 
+from .events import *
 from .logs import *
+from .merkle import *
+from .pending_transfers import *
+from .proofs import *
+from .sign import *
+from .types import *
