+++ raiden_contracts/utils/pending_transfers.py
@@ -1,45 +1,23 @@
 from random import randint
-from os import urandom
 from collections import namedtuple
 from functools import reduce
 
 from eth_abi import encode_abi
-from web3 import Web3
-from web3.utils.threads import (
-    Timeout,
-)
 
 from raiden_contracts.constants import TEST_SETTLE_TIMEOUT_MIN
 from raiden_contracts.utils.merkle import compute_merkle_tree, get_merkle_root
+from raiden_contracts.tests.utils.mock import random_secret, get_random_values_for_sum
 
 
-def check_succesful_tx(web3: Web3, txid: str, timeout=180) -> dict:
-    '''See if transaction went through (Solidity code did not throw).
-    :return: Transaction receipt
-    '''
-    receipt = wait_for_transaction_receipt(web3, txid, timeout=timeout)
-    txinfo = web3.eth.getTransaction(txid)
-    assert receipt['status'] != 0
-    assert txinfo['gas'] != receipt['gasUsed']
-    return receipt
-
-
-def wait_for_transaction_receipt(web3, txid, timeout=180):
-    with Timeout(timeout) as time:
-            while not web3.eth.getTransactionReceipt(txid):
-                time.sleep(5)
-
-    return web3.eth.getTransactionReceipt(txid)
-
-
-def get_random_values_for_sum(values_sum):
-    amount = 0
-    values = []
-    while amount < values_sum:
-        value = randint(1, values_sum - amount)
-        values.append(value)
-        amount += value
-    return values
+PendingTransfersTree = namedtuple('PendingTransfersTree', [
+    'transfers',
+    'unlockable',
+    'expired',
+    'packed_transfers',
+    'merkle_tree',
+    'merkle_root',
+    'locked_amount',
+])
 
 
 def get_pending_transfers_tree(
@@ -130,19 +108,3 @@ def get_packed_transfers(pending_transfers, types):
 
 def get_locked_amount(pending_transfers):
     return reduce((lambda x, y: x + y[1]), pending_transfers, 0)
-
-
-PendingTransfersTree = namedtuple('PendingTransfersTree', [
-    'transfers',
-    'unlockable',
-    'expired',
-    'packed_transfers',
-    'merkle_tree',
-    'merkle_root',
-    'locked_amount',
-])
-
-
-def random_secret():
-    secret = urandom(32)
-    return (Web3.soliditySha3(['bytes32'], [secret]), secret)
